refactor (websocket): sending message
