+++ constant.ts
@@ -1,5 +1,7 @@
 const BYTE_CONSTANT =  {
-    FIRST_BIT_VALUE: 128
+    FIRST_BIT_VALUE: 128,
+    MAXIMUM_SIXTEEN_BITS_VALUE: 2 ** 16,
+    MAXIMUM_SIXTY_FOUR_BITS_VALUE: 8 ** 16,
 }
 
 export {
