+++ enum.ts
@@ -3,6 +3,9 @@ const SocketEnum = {
             sevenBits : 125,
             sixteenBits: 126,
             sixtyFourBits: 127,
+        },
+        opCode: {
+            text: 0x01,
         }
 }
